spline #478 Pre-release refactoring: Persistence layer - re-arrange persistent model source files
