+++ persistence/src/main/scala/za/co/absa/spline/persistence/model/typeDefs.scala
